+++ core/envelope-sim/src/main/java/fr/sncf/osrd/envelope_sim/power/EnergySourceType.java
@@ -1,0 +1,7 @@
+package fr.sncf.osrd.envelope_sim.power;
+
+public enum EnergySourceType {
+    CATENARY,
+    POWER_PACK,
+    BATTERY
+}
