+++ core/envelope-sim/src/main/java/fr/sncf/osrd/envelope_sim/power/storage/ManagementSystem.java
@@ -1,0 +1,11 @@
+package fr.sncf.osrd.envelope_sim.power.storage;
+
+public class ManagementSystem{
+    double overchargeThreshold;          //overcharge limit
+    double underchargeThreshold;         //undercharge limit
+
+    public ManagementSystem(double overchargeThreshold, double underchargeThreshold) {
+        this.overchargeThreshold = overchargeThreshold;
+        this.underchargeThreshold = underchargeThreshold;
+    }
+}
